refactor(scripts): decompose post-merge-pipeline into phases/
Extracts the six post-merge phases plus their private helpers from the
monolithic post-merge-pipeline.js (862 LOC) into a sibling phases/
directory under .agents/scripts/lib/orchestration/post-merge/:

  - worktree-reap.js         (395 LOC) + private Windows-lock helpers
  - ticket-closure.js        (107 LOC)
  - notification.js          ( 78 LOC)
  - temp-cleanup.js          ( 68 LOC)
  - branch-cleanup.js        ( 56 LOC)
  - dashboard-refresh.js     ( 33 LOC)

The parent post-merge-pipeline.js shrinks to 205 LOC and becomes a
thin sequencer (runPostMergePipeline + DEFAULT_POST_MERGE_PHASES +
perfSummaryPhase) that re-exports every extracted phase, preserving
the existing import surface so call sites and tests are unchanged.

Pure code-move refactor; no behavior change. 6751 tests pass.

Refs #3000 #2988
